geolocation/geo-getCurrentPosition-004.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: call getCurrentPosition with options maximumAge set string, then get Coordinates object and no exception expected</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#get-current-position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var t =async_test("Test checks that call getCurrentPosition with options maximumAge set string, then get Coordinates object and no exception expected");
      setup({timeout: 20000});
      try {
        window.navigator.geolocation.getCurrentPosition(successCallback, errorCallback, {maximumAge: "tests"});
      } catch (err) {
        t.step(function () {
          assert_unreached("Error message: " + err.message);
        });
        t.done();
      }
      function successCallback (position) {
        t.step(function () {
          assert_equals(typeof position.coords, "object", "call getCurrentPosition with options maximumAge set string, then get Coordinates object and no exception expected");
        });
        t.done();
      }
      function errorCallback (error) {
        t.step(function () {
          assert_unreached("Error message: " + error.message);
        });
        t.done();
      }
    </script>
  </body>
</html>
